test: ensure gRPC gateway security config is consistent with REST
